refactor: extract qcow2 header types into meta/header.rs
Move Qcow2RawHeader, the header-extension types, the feature-type enum
and Qcow2Header (with its parse/format/serialize logic) into a new
meta/header.rs submodule, re-exported from mod.rs. The remaining table
types stay in mod.rs and header.rs pulls the refcount table types it
needs for format_qcow2 from super. No logic changes.
